+++ 6502sharp/instructions/BEQ.cs
@@ -10,7 +10,7 @@ namespace _6502sharp.Instructions
         [CPUInstruction(0xF0, 2)]
         public void BEQ_Relative([Relative] int target)
         {
-            if (cpu.SR.Zero == true) cpu.PC.Value = target;
+            if (cpu.SR.Zero == true) cpu.PC.Value = (ushort)target;
         }
     }
 }
