+++ 6502sharp/instructions/BNE.cs
@@ -10,7 +10,7 @@ namespace _6502sharp.Instructions
         [CPUInstruction(0xD0, 2)]
         public void BNE_Relative([Relative] int target)
         {
-            if (cpu.SR.Zero == false) cpu.PC.Value = target;
+            if (cpu.SR.Zero == false) cpu.PC.Value = (ushort)target;
         }
     }
 }
