+++ 6502sharp/instructions/RTS.cs
@@ -12,7 +12,7 @@ namespace _6502sharp.Instructions
         [CPUInstruction(0x60, 6)]
         public void RTS_Implied()
         {
-            cpu.PC.Value = cpu.Stack.PopPC();
+            cpu.PC.Value = (ushort)cpu.Stack.PopPC();
             cpu.PC.Value++;
         }
     }
